app: moved response transform lib
