+++ src/libs/common/transform/transform.decorator.ts
@@ -2,6 +2,6 @@ import { SetMetadata } from "@nestjs/common"
 
 export const TRANSFORM_RESPONSE_DTO_KEY = "transform_response_dto"
 
-export function TransformResponse(dto: any) {
+export function TransformResponse(dto: unknown) {
 	return SetMetadata(TRANSFORM_RESPONSE_DTO_KEY, dto)
 }
